requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==4.2.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0